+++ src/settings/Plugins/Plugins.js
@@ -3,7 +3,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { modules } from 'stripes-config'; // eslint-disable-line import/no-unresolved, import/no-extraneous-dependencies
-import { Callout } from '@folio/stripes/components';
+import {
+  Callout,
+  Layout,
+} from '@folio/stripes/components';
 import PluginForm from './PluginForm';
 
 class Plugins extends React.Component {
@@ -103,7 +106,7 @@ class Plugins extends React.Component {
   render() {
     const plugins = this.getPlugins();
     return (
-      <div style={{ width: '100%' }}>
+      <Layout className="full">
         <PluginForm
           onSubmit={this.save}
           label={this.props.label}
@@ -111,7 +114,7 @@ class Plugins extends React.Component {
           initialValues={{ plugins }}
         />
         <Callout ref={(ref) => { this.callout = ref; }} />
-      </div>
+      </Layout>
     );
   }
 }