+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [UITEN-85](https://issues.folio.org/browse/UITEN-85) Settings > Tenant Pages > Move Save button to the Footer - Implement the pane footer component
 
 ### Bugs
+* [UITEN-32](https://issues.folio.org/browse/UITEN-32) Remove hardcoded font sizes
 * [UITEN-72](https://issues.folio.org/browse/UITEN-72) Security update eslint to >= 6.2.1 or eslint-util >= 1.4.1
 
 ## [3.0.0](https://github.com/folio-org/ui-organization/tree/v3.0.0) (2020-03-12)